+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/kenchan0130/twitter-like-rss
+module github.com/kenchan0130/twitter-like-feed
 
 go 1.14
 
@@ -7,4 +7,6 @@ require (
 	github.com/gin-gonic/gin v1.6.3
 	github.com/gorilla/feeds v1.1.1
 	github.com/kr/pretty v0.2.0 // indirect
+	github.com/mmcdole/gofeed v1.0.0
+	github.com/stretchr/testify v1.6.1
 )
